Updated to used configurable WiFi Credentials
Added the WiFiManager library. Refactored code so that now on first boot the ESP will create it's own WiFi Client Server where you can log on and add the users WiFi network credentials.

Corrected the data type of the global constants declared at initialisation.

Refactored/renamed the "ESP Scheduler" so that it no longer conflicts with the standard Arduino "Scheduler". The library was also refactored to ensure compatibility. i intend to contribute to the open source code to save others this trouble in future.